Switch to async Ingress Reading (#1180)
* Async read ingress
* split up ingress.go
